Xong

+++ src/main/java/tad/controller/employee/EmployeeController.java
@@ -62,7 +62,7 @@ public class EmployeeController {
 	public String gInfo(ModelMap model, HttpSession session) {
 		Account acc = (Account) session.getAttribute(DefineAttribute.UserAttribute);
 		if (acc == null) {
-			return "redirect:/";
+			return "redirect:/guest.htm";
 		}
 
 		UserBean userBean = new UserBean(acc.getEmail(), acc.getFirstName(), acc.getLastName(), acc.getPhoneNumber());
@@ -76,7 +76,7 @@ public class EmployeeController {
 			BindingResult errors, HttpSession session, ModelMap modelMap) {
 		Account acc = (Account) session.getAttribute(DefineAttribute.UserAttribute);
 		if (acc == null) {
-			return "redirect:/";
+			return "redirect:/guest.htm";
 		}
 
 		if (errors.hasErrors()) {
@@ -94,25 +94,27 @@ public class EmployeeController {
 		acc.setPhoneNumber(user.getPhoneNumber());
 		acc.setEmail(user.getEmail());
 
-		File file = new File(rootFile.getPath() + user.getAvatar());
-		if (file.exists())
-			file.delete();
+		if (user.getAvatar().isEmpty()) {
+		} else {
+			File file = new File(rootFile.getPath() + user.getAvatar());
+			if (file.exists())
+				file.delete();
 
-		String avatarPath = accountImgDir.getPath() + user.getAvatar();
-		acc.setAvatar(user.getAvatar().getOriginalFilename());
+			String avatarPath = accountImgDir.getPath() + user.getAvatar();
+			acc.setAvatar(user.getAvatar().getOriginalFilename());
 
-		try {
-			user.getAvatar().transferTo(new File(avatarPath));
-			Thread.sleep(2000);
-		} catch (Exception e) {
-			e.printStackTrace();
-			modelMap.addAttribute("message", false);
-			modelMap.addAttribute(DefineAttribute.UserBeanAttribute, user);
-			return "employee/employee-profile";
+			try {
+				user.getAvatar().transferTo(new File(avatarPath));
+				Thread.sleep(2000);
+			} catch (Exception e) {
+				e.printStackTrace();
+				modelMap.addAttribute("message", false);
+				modelMap.addAttribute(DefineAttribute.UserBeanAttribute, user);
+				return "employee/employee-profile";
+			}
 		}
 
 		accountDAO.updateAccount(acc);
-
 		modelMap.addAttribute(DefineAttribute.UserBeanAttribute, user);
 		modelMap.addAttribute("message", true);
 		return "employee/employee-profile";
